requirements.txt
black==21.8b0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
fiql-parser==0.15
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
idna==3.2
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.3.0
powerschool==2.0.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
